[iOS] Refactor: 기존의 데이터소스 객체를 없애고, 뷰컨이 데이터소스 역할을 하도록 기능 구현함 	이유: 일단 역할을 뷰컨이 모두 하게 하고 이후에 따로 데이터소스, 델리게이트 객체를 만들어서 	      역할을 분리하도록 할 것이다.
